app/api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
spacy>=3.7.0
openai>=1.12.0
litellm>=1.30.0
supabase>=2.4.0
pydantic>=2.6.0
pandas>=2.2.0
python-dotenv>=1.0.0
httpx>=0.26.0
cryptography>=42.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-multipart>=0.0.9
google-genai>=0.2.0
instructor>=1.0.0
